requirements.txt
fade
colorama
phonenumbers
pystyle
gitpython
discord
discord.py
opencv-python
pillow
requests
pyshorteners